Update: Make AcitiveMarker and InfoWindow a component
