+++ components/PlaceInfo/index.js
@@ -1,13 +1,9 @@
 /* eslint-disable @next/next/no-img-element */
 import React, { useContext, useEffect } from "react";
-import { MarkerF, InfoWindowF } from "@react-google-maps/api";
-import { IconContext } from "react-icons";
-import { BiLinkExternal } from "react-icons/bi";
-import { BsStarFill, BsFillHouseHeartFill, BsPhoneFill } from "react-icons/bs";
-import styles from "./index.module.css";
+import { MarkerF } from "@react-google-maps/api";
 import AppContext from "@/context/AppContext";
-import Image from "next/image";
-import FavoriteButton from "../FavoriteButton";
+import ActiveMarker from "./ActiveMarker";
+import InfoWindow from "./InfoWindow";
 
 export default function PlaceInfo({ businessList, isListView }) {
   const {
@@ -25,11 +21,6 @@ export default function PlaceInfo({ businessList, isListView }) {
       return;
     }
     setSelected(marker);
-    setSelectedIndex(null);
-  };
-
-  const handleCloseInfoWindow = () => {
-    setSelected(null);
     setSelectedIndex(null);
   };
 
@@ -65,82 +56,8 @@ export default function PlaceInfo({ businessList, isListView }) {
           }}
         />
       ))}
-
-      {selectedIndex && (
-        <MarkerF
-          key={selectedIndex}
-          position={{
-            lat: businessList[selectedIndex - 1]?.coordinates.latitude,
-            lng: businessList[selectedIndex - 1]?.coordinates.longitude,
-          }}
-          label={{
-            text: `${selectedIndex}`,
-            color: "#ff0000",
-          }}
-          icon={{
-            path: google.maps.SymbolPath.CIRCLE,
-            fillColor: "#fff",
-            fillOpacity: 1,
-            scale: 14,
-            strokeColor: "#ff0000",
-            strokeWeight: 1,
-          }}
-          zIndex={999}
-        />
-      )}
-
-      {selected && (
-        <InfoWindowF
-          position={{
-            lat: selected.coordinates.latitude,
-            lng: selected.coordinates.longitude,
-          }}
-          onCloseClick={handleCloseInfoWindow}
-        >
-          <div className={styles.infoWindows}>
-            {selected.image_url && (
-              <img
-                src={selected.image_url}
-                alt="Restaurant image"
-                className={styles.image}
-              />
-            )}
-            {!selected.image_url && (
-              <Image
-                src="/image/No_Image_Available.jpg"
-                alt="No image available"
-                width={150}
-                height={150}
-                className={`${styles.image} ${styles.image__noimage}`}
-              />
-            )}
-            <h1>{selected.name}</h1>
-            <p className={styles.alias}>{selected.alias}</p>
-            <IconContext.Provider value={{ className: styles.star }}>
-              <p>
-                <BsStarFill /> {selected.rating}
-              </p>
-            </IconContext.Provider>
-            <p className={styles.categories}>{selected.categories}</p>
-            <p className={styles.location}>
-              <BsFillHouseHeartFill className={styles.icon} />{" "}
-              {selected.location}
-            </p>
-            <p className={styles.phone}>
-              <BsPhoneFill className={styles.icon} /> {selected.phone}
-            </p>
-
-            <div className={styles.footer}>
-              <IconContext.Provider value={{ className: styles.link }}>
-                <a href={selected.url} target="blank" rel="noopener noreferrer">
-                  Link <BiLinkExternal />
-                </a>
-              </IconContext.Provider>
-              <FavoriteButton businessListItem={selected} />
-            </div>
-          </div>
-        </InfoWindowF>
-      )}
+      {selectedIndex && <ActiveMarker businessList={businessList} />}
+      {selected && <InfoWindow />}
     </>
   );
 }
